address

+++ Kooboo.Sites/Ecommerce/KScript/KOrder.cs
@@ -25,52 +25,64 @@ namespace Kooboo.Sites.Ecommerce.KScript
         }
 
         [Description("Create an order based on current shopping cart items")]
-        public Order Create()
+        public Order Create(string addressId)
         {
-            var cartservice = ServiceProvider.GetService<CartService>(this.context);
-            var cart = cartservice.GetOrCreateCart();
-            if (cart.Items.Any())
+            Guid id;
+            bool parseok = Guid.TryParse(addressId, out id);
+
+            if (parseok)
             {
-                var order = this.service.CreateOrder(cart);
-                return order;
+                var cartservice = ServiceProvider.GetService<CartService>(this.context);
+                var cart = cartservice.GetOrCreateCart();
+                if (cart.Items.Any())
+                {
+                    var order = this.service.CreateOrder(cart, id);
+                    return order;
+                }
             }
             return null;
         }
-         
-        public Order CreateSelected(object[] selected)
-        {
-            List<Guid> selectedCartItem = new List<Guid>();
-            foreach (var item in selected)
-            {
-                if (item !=null)
-                {
-                    var stritem = item.ToString();
-                    
-                    if (System.Guid.TryParse(stritem, out Guid id))
-                    {
-                        selectedCartItem.Add(id); 
-                    }
-                }
-            }
 
-            var cartservice = ServiceProvider.GetService<CartService>(this.context);
-            var cart = cartservice.GetOrCreateCart();
-            if (cart.Items.Any() && selectedCartItem.Any())
-            { 
-                var order = this.service.CreateOrder(cart.Items.FindAll(o=>selectedCartItem.Contains(o.Id)).ToList());
-                if(order!=null)
+        public Order CreateSelected(object[] selected, string addressId)
+        {
+            Guid addressguid;
+            bool parseok = Guid.TryParse(addressId, out addressguid);
+
+            if (parseok)
+            {
+                List<Guid> selectedCartItem = new List<Guid>();
+                foreach (var item in selected)
                 {
-                    foreach (var item in order.Items)
+                    if (item != null)
                     {
-                        cartservice.RemoveItem(item.ProductVariantId);
+                        var stritem = item.ToString();
+
+                        if (System.Guid.TryParse(stritem, out Guid id))
+                        {
+                            selectedCartItem.Add(id);
+                        }
                     }
                 }
-                return order;
+
+                var cartservice = ServiceProvider.GetService<CartService>(this.context);
+                var cart = cartservice.GetOrCreateCart();
+                if (cart.Items.Any() && selectedCartItem.Any())
+                {
+                    var order = this.service.CreateOrder(cart.Items.FindAll(o => selectedCartItem.Contains(o.Id)).ToList(), addressguid);
+                    if (order != null)
+                    {
+                        foreach (var item in order.Items)
+                        {
+                            cartservice.RemoveItem(item.ProductVariantId);
+                        }
+                    }
+                    return order;
+                }
             }
 
             return null;
         }
-         
+
         public void Paid(object[] OrderIds)
         {
             foreach (var OrderId in OrderIds)
@@ -96,9 +108,4 @@ namespace Kooboo.Sites.Ecommerce.KScript
             return orders;
         }
     }
-
-
-
-
-
 }
